Corrected required width/length from float to int based on provided test cases.

+++ swap_meet/decor.py
@@ -3,8 +3,8 @@ from .item import Item
 class Decor(Item):
     def __init__(self, year=None, id=None, width=0, length=0, condition=0):
         super().__init__(year=year, id=id, condition=condition)
-        self.width = width if width==0 else float(width)
-        self.length = length if length==0 else float(length)
+        self.width = width if width==0 else int(width)
+        self.length = length if length==0 else int(length)
         if width < 0 or length < 0:
             raise ValueError
